requirements.txt
requests
PySocks
fake-useragent
colorama
configparser